oci: add oci image handling
Retrieve a base image from a docker registry, extract it to use as a
craft-parts overlay base image, repack the parts prime content as
a new layer and make the image available to the local docker daemon.

Other final artifacts can be made available in the future, such as a
exported docker image archive or OCI image archive.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>