fix: --preserve-digests leads to unsupported manifest media type
When using --preserve-digests with Skopeo, the final OCI artefact will have the right digest,
but the manifest type may not be supported.

For example, when creating image from Docker registry, with --preserve-digests,
the corresponding OCI image will have a manifest with the media type
"application/vnd.docker.distribution.manifest.v2+json", which IS NOT
supported by umoci.

Instead, we must always get the digest directly from the source image.