oci: preserve directories when adding layers
The previous code walked() on the source dir "manually" adding the
files, possibly to preserve a path relative to the source dir. This had
the side effect of not explicitly adding the sub directories, which
meant that their attributes (permission, ownership) were not preserved
in the final image (as they couldn't be!). So we simplify to add the
whole directory at once, with a proper `arcname` to preserve the
relative paths.

Fixes #44