spread: improve 'prepare' and 'restore'
During 'prepare' we use 'retry' to be a bit more lenient when waiting
for the Docker daemon to start, because it's sometimes failing with the
previous default of sleeping for 5 seconds.

In 'restore-each' we prune the docker images and cleanup lxc instances,
to avoid blowing up disk usage between tasks.