parts: preserve existing "pebble" part
This provides a way to "customize" pebble inside the ROCKs: if a
project's spec already contains a "pebble" part, we preserve it and
assume it's working. This is done in load_project(), _before_
unmarshaling, to be consistent with what Snapcraft does for similar
"special" parts.

The new _add_pebble_data() will also be useful in the future when we
have to configure entrypoints, pebble env variables, etc.