oci: handle symlinks in base layer (usrmerge)
This commit is a first step in gracefully handling existing layers when
packing the parts' contents. Specifically, it fixes the existing
breakage where priming a file in "/lib" will break a resulting
Ubuntu-based rock because on the base "/lib" is a symlink to "/usr/lib",
and if that link is broken everything breaks.

The updates to support this are:

- Enhance oci.Image.add_layer() to take the existing base into account,
  via a parameter that points to the extracted base (see below);
- Update lifecycle code to pass the extracted base forward when packing;
- Add a jammy runner to CI, because the new integrated tests need umoci
  which is only available on jammy.

The update to add_layer() is to look for a file's parent dir in the
existing base. If it *does* exist, and is a symlink to another dir, add
the file as belonging to the symlink target instead. For example:

- The prime dir contains "bin/file.txt";
- The base is ubuntu:22.04, which means that there "/bin/" is a symlink
  to "/usr/bin";
- Instead of adding the file as "bin/file.txt", add it as
  "usr/bin/file.txt".