spread: retry rockcraft commands in fedora (#179)
* spread: move from fedora-35 to fedora-36

Fedora 35 is EOL and its main package repo got archived today.

* spread: retry rockcraft commands in fedora

Because of a snapd 2.58 bug[1], rockcraft may fail when running "snapd
set ..." commands inside the instance. This only happens once, and the
command succeeds the second time. Add a wrapper around rockcraft calls
to retry when in Fedora while the bug is not fixed.

[1]: (https://bugs.launchpad.net/snapd/+bug/2002835

* disable tutorials/code spread on fedora