oci: use alternative registry
Docker Hub is fast but the Ubuntu images are updated infrequently and
the registry itself has pull limits for anonymous users. This limit is
annoying because currently rockcraft does not caching of base images.

The caching is coming, but in the meantime switch to Amazon ECR. The
registry has newer images and is only a little bit slower (from testing)
than Docker Hub).