oci: fix handling of subdirs in usrmerge (#200)
The scenario that was broken was: a payload like "bin/dir1/dir2" where
"bin" is a symlink to somewhere on the lower (base) layer. The subdirs
"dir1" and "dir2" were being incorrectly placed in "bin" instead of in
the symlink target.

This broke quite a few plugins that place file trees in places like
"lib/".