oci: handle duplicate directories because of usrmerge
The situation is this: the primed payload has two distinct directories
(e.g. bin/dir1 and usr/bin/dir1) that, because of the usrmerge handling,
"point" to the same location (in this case, /usr/bin/dir1). We check
that the two directories have the same attributes (ownership and mode)
and then add a single entry to the layer (either one, they are all
equivalent).

All other cases (dirs with different attributes, conflicts between
files) raise errors. The reason for this is that we haven't seen this
come up "in the wild" yet, and it's hard to choose a behavior when we
don't know the circunstances.

Fixes #203