oci: support file collisions if the files match (#231)
In c3382fc2408 we added support for directory collisions caused by the
usrmerge when creating the payload layer; now we add support for file
collisions if the files match in attributes and contents. This scenario
can happen if an Ubuntu package is used both as an overlay-package and
as a stage-package.