plugins: add sitecustomize.py in PythonPlugin's build (#248)
The issue: running a ROCK that was built with the Python plugin by
calling "python3" or "/usr/bin/python3" (as opposed to "/bin/python3")
fails to find pip-installed packages.

The solution implemented with this commit is to add a sitecustomize
module that always adds the location of the pip-installed packages to
sys.path. This ensures that the pip-installed packages are found even
if the "global" Python interpreter is called.