spread: retry grepping docker logs (#282)
One possible source of flakiness in spread tests is the common idiom
of starting a container and then immediately grepping the logs. Since
our containers' entrypoint is Pebble, it might take a short amount of
time to startup, create the services plan, and start whatever it is
we want it to start.

Therefore, add a helper command to 'retry' this 'docker logs | grep'
pipe.