parts: add project-specific global variables
The implementation follows Snapcraft's: a prologue callback is used to
fill the project info with the project's name and version (the other
'useful' variables are already provided by craft_parts). This lets
parts use those variables (for example during their builds).

A lot of the unit test changes come from a refactoring to make adding
new parameters to parts.PartsLifecycle less painful in the future.

Closes #299