replace 'black' with a simple hello script
'black' depends on a package (typed-ast) that had a recent release that,
for some reason, doesn't have wheels for the build's setup. This means
it tries to build the package, which has C code, which then fails
because we're not including dev dependencies.

Since the point of the test is just to check that a ROCK with python 3.6
can run OK, remove black and add a simple script in its place.