feat: add image service
This service does the fetching/converting of OCI images for the use
of the lifecycle and packaging services. It also has a very simple
"cache" since it only fetches things once.

The test is lacking because the majority of the code was copied over
from rockcraft.lifecycle. We can/should improve this in the future.