build: set version from git
This commit pulls in the "dynamic" version setting (from git) from
Starbase, and updates snapcraft.yaml to use that version (via
adopt-info).

A consequence of this is that, in order to be able to run "python3
setup.py --version" without the dependencies installed, the enabling of
the overlay feature is moved from ``rockcraft/__init__.py`` to the
lifecycle service, which makes more sense any way as it's a feature
specific to that service.