feat: prune primed files if they exist on the base
This commit adds a lifecycle callback that "prunes" (removes) those
primed files that already exist (with same contents and permissions) in
the base layer. This is necessary because the dependency resolver
doesn't look into the base layer when deciding dependency cutoffs (and
also to support chisel slices, which only look at the target install
dir). Some notes:

1) This first implementation only looks are regular files (no symlinks
and directories), as those are the simplest to handle and provide the
biggest size "wins". If necessary, we can address symlinks and/or
directories in follow-up work.
2) This could have been implemented during the archiving of the new
layer, where we already loop on the files and look at the base layer (to
handle things like the usrmerge). We decided to implement this as part
of the lifecycle to improve the "clarity" and "debuggability" of
Rockcraft runs: after running the lifecycle, the prime directory
contains exactly what's going to be packed in the new layer.

The resulting ROCK should be "semantically" equivalent, but smaller. For
instance, the Pyfiglet ROCK in docs/tutorials/code/pyfiglet goes from
69MB to 53MB, with no functional changes.

Fixes #363