fix: use correct error message for invalid version
The code refers to the 'name' message, but the bug was due to the
"error_msg_templates" in the Config: it made *every* regex validation
failure use the INVALID_NAME_MESSAGE message, even for those fields that
*aren't* the 'name'.

Since the "error_msg_templates" option is deprecated and gone in
Pydantic 2 anyway, remove it altogether and use the new
transform_pydantic_error() hook from craft-application to preserve the
customisation of the 'name' field's message. This is necessary because
our 'name' regex is different than the 'name' regex in
craft-application's model (ours is a bit stricter), so we can't just use
craft-application's message as that would be misleading.

Fixes #429