fix(pebble): /var/lib/pebble/default has mode 0777 (#473)
This commit ensures that the PEBBLE default directory
/var/lib/pebble/default always has the mode 0777.

Previously, the mode of the PEBBLE dir was only set to 0777 if there
were services or checks specified in the rockcraft.yaml. Otherwise, it
used to be 0755 with user=root group=root. This meant that if pebble was
being run as some user other than root, using the `run-user` directive
or some other way, the pebble run would fail. It would fail because
pebble could not create the necessary files (socket) inside that
directory. This bug is reported in issue #461.

Fixes #461