build(deps): update skopeo to 1.14.2 (#477)
Version 1.14.2 fixes the issue with Docker Daemon API mismatch when converting the rock to a Docker image. The previous version (1.9.0) is also over a year and a half old so it's a good idea to update anyway.

This also includes normalizing the go version used to build the go dependencies (skopeo and umoci) to 1.21.