feat: add the `flask-framework` extension (#499)
This commit is the squashed version of the long-lived `feature/12f` branch containing the most recent version of the Flask extension in the context of the 12-factor App methodology. The original commit log is kept here for posterity, and follows:

------

* feat: new flask extension (#317)

Co-authored-by: Sergio Schvezov <sergio.schvezov@canonical.com>
Co-authored-by: Tiago Nobrega <tiago.nobrega@canonical.com>

* feat: provide a default prime in the flask extension (#374)

* feat(extensions): add a profile option to the rockcraft init command (#379)

* chore(deps): update craft-providers to 1.19.2

The current version has two bugs (the "looping timer" one and the "not
collecting error logs" one) that are blocking for many users. Ideally we
should merge main into this branch, this is a stopgap before we can
allocate time for it.

* chore: manual updates to align with origin/main

These are related to the change in base notation (ubuntu@ instead of
ubuntu:) and the changes due to general craft-application-flow.

* fix(docs): remove duplicate how-to entries

* Fix schema validation

* Fix typo

* Template requires input variables

* Fix lint

* Fix lint

* Update flask-framework extension

* Remove support for ubuntu:20.04

* ROCK to rock and update copyright info

* Update flask doc with real examples

* Update use-flask-extension.rst

* Update pyright exclude

* Fix some linting issues

* enable extension in spread tests

* Fix use-flask-extension spread test

* Rename ROCK to rock

* Apply suggestions from code review

Co-authored-by: Alex Lowe <alex.lowe@canonical.com>

---------

Co-authored-by: Sergio Schvezov <sergio.schvezov@canonical.com>
Co-authored-by: Tiago Nobrega <tiago.nobrega@canonical.com>
Co-authored-by: Gregory Schiano <gregory.schiano@canonical.com>
Co-authored-by: Alex Lowe <alex.lowe@canonical.com>