refactor(docs): use more docs from craft-parts
Now that the structure of the shared craft-parts docs has more content,
we can simplify our own docs and just use craft-parts' directly in many
places.

The guideline is this: If the document will be used wholesale, just reference
it directly in the TOC. Otherwise, `include` it in a Rockcraft doc.