fix(spread): more robust check of non-pruned files
The spread test for the "layer pruning" feature worked by comparing the stats
of two files if they exist in both layers. These stats - filesize, ownership
and permission bits - were enough up until now because the non-pruned files
differed on at least the filesize.

The recent release of libuuid1 broke this expectation: this new package has a
``libuuid.so.1.3.0`` with identical file size as the previous version, but
different contents. So the pruning correctly kept the file, but the test failed.

Update the test to _also_ check the files' contents if necessary.