spread: reorganize spread tests (#519)
Reorganize the spread tests to make better use of CI resources (Google VMs).
The majority of the tests work the same regardless of host system, so it's a
waste to run them on jammy _and_ focal _and_ fedora. Instead, bump the number
of jammy machines and have them run the majority of the tests (renamed from
`general` to `rockcraft'`), and add a new suite `all-systems` where we can add
tests that are explicitly designed to test cross-system support.