chore: force build-base and platform on flask examples
The problem is this: the flask extension currently adds a 'build-base' if the
'base' is 'bare' but the 'build-base' is undeclared, and a 'platforms' entry
if it's missing,  but this behavior breaks with the new version of
craft-application because the BuildPlanner processes the yaml _before_ the
extensions are applied, and the planner cannot work without a build-base or
platform description.

The extension behavior of adding a 'build-base' and a 'platforms' item needs
to be reviewed: we probably don't want to support this. In the meantime,
this commit fixes only the projects that are executed in spread.