fix(pebble): place pebble in a separate location (#528)
The context here is that starting from Ubuntu 24.04, the 'base-files' package
(and related chisel slices) provides "bin" as a symlink to "usr/bin". This
breaks the previous "phantom" pebble part because it created "bin" as a regular
directory which then conflicts with the symlink.

Moving forward, the pebble binary is now placed in ".rock/bin/". This way we
won't get further collisions and this reflects the fact that the location of
the binary is, and should be seen, as an implementation detail.