fix(tests): don't use chisel slices in 'devel' bases
The issue is that chisel releases necessarily lag behind mainline Ubuntu
development, as each new series needs to be manually added to the
chisel-releases repo. The case in point here is that 'devel' has just
started to mean "Ubuntu 24.10", but this version is not available in
chisel-releases yet.

So we drop the use of slices in 'devel' tests, which should be fine because
the point is to see whether the *rest* of Rockcraft's execution works.