docs: use "dev" for non-release builds
The bug is this: If a build is done off of (say) 'main', the version coming
from 'git describe' will be something like 1.32.2.postXXXX. We previously
dropped the ".post*" suffix because it confused the spellchecker but that
just leaves the misleading "1.32.2" string. It's misleading because that's
an actual released version that does *not* contain a lot of the content
in 'main'.

So just use "dev" instead, to make it extra explicit that the docs refer
to an unreleased version of the software.