chore(ci): move base devel spread tests to manual
These tests are by their very nature flaky because the base devel changes
constantly and can occasionally break in a way that cannot be fixed by us.
However, we still want to keep an eye on them so repurpose the 'large'
spread run into a 'manual' one that, via cron, runs both the 'large' tests
and the 'flaky' ones.