docs: use more inclusive language in flask tutorial (#707)
I discussed with tutorial with Daniele, who expressed some concerns about the way the tutorial was written. Emphasizing the user doing something ("you do X, then you do Y, ...") gives the impression of a how-to guide. A tutorial is more about doing something together with the user as a learning experience.

Therefore, I changed some of the language from "you" to "we" to frame the experience as more of a tutorial. We're doing this together with the user so that they can learn how to build a rock for a Flask app. 

I left some instances of "you" alone where I thought it was appropriate to include.

---------

Co-authored-by: Michael DuBelko <michael.dubelko@gmail.com>