feat: add Rockcraft-specific Poetry plugin
The plugin has the same behavior and restrictions as the Python plugin with
regards to base-dependent behavior, symlink handling, sitecustomize, etc.

Therefore, this common behavior is extracted into a new "python_common"
module, used by both plugins. This approach is also taken for the reference
docs - the requirement of staging a Python interpreter is the same for both
plugins.

Fixes #701