build(deps): update deps with freeze-requirements
This commit also updates freeze-requirements to use 'pip-compile' instead of
'pip freeze'. The reason for this is that 'pip freeze' does not preserve the
'extras' of each dependency listed in the project's dependencies, which in
this particular case is relevant for the 'canonical-sphinx[full]' change.

pip-compile also has the side-effect of showing where each transient dependency
came from, which is very nice!