fix: accept a 'pebble' part if its the same as ours
The use-case that this enables is packing the output of 'expand-extensions'.
Previously, this would fail because the expanded yaml contains the 'pebble'
part that Rockcraft adds, and then Rockcraft itself would complain that the
project already has a part called 'pebble'.

This commit makes Rockcraft accept an existing part called 'pebble' if the
contents of said part are the same as the one we would add.

Fixes #749