ci(spread): only run docs tests if docs are changed
These tests take a disproportional amount of time to run, and don't need to run
on every code change. This adds a separate job, 'docs-tests', that runs the
spread tests for the docs but only if:

a) the pull request has the 'documentation' label, or
b) the pull request has the 'run-manual-spread' label, or
c) this is a scheduled run or
d) this is a run in 'main'.

Note that the 'documentation' label can either be added manually to a PR, or
automatically if the changeset includes files in "docs/" (see the labeler job).