build(deps): update dependencies
This includes fixing an expected-message test. The previous version was
temporary, and a new craft-platforms update brings in the correct, expected UI
of "Debian architecture".

This also includes adding a cross-compilation test (from the craft-archives
update). This test is analogous to Snapcraft's, and builds a riscv64 rock with
a binary created via cross-compilation.