refactor: use craft-application 5 (#856)
* build: pin python-apt for oracular

* build: use craft-app@main

* feat: add a project service

* refactor: fix immediate issues with craft-app 5

* wip: checkpoint, snap is packable

* wip: checkpoint

* fix(test): fix the expand-environment and pebble tests on the app

* wip: checkpoint

* test: use fake_project_file fixture

* test(ProjectService): switch unit tests from application

This moves unit tests that were previously being done on the application
into the ProjectService which actually does it.

* test(PackageService): use real platforms for the package service

* test: add customizable fake_project_file overrides

* test: fix package test with new fake project file fixture

* test: adjust fixtures for lifecycle package repo test

* test: fix extension tests

* test: fix python plugin integration tests

* fix(test_oci): use new infrastructure for oci tests

* fix(tests): fix Python plugin integration tests

* test(oci): fix oci test for the fact that we're adding pebble

* test: fix one xfailing test, remove another that's upstreamed now

* fix: set kwargs for image service

* chore: delete now-unused testing file

* fix(test): fix lifecycle integration tests

* chore(spread): make 'restore' scripts more robust

A restore script that fails to run has the side effect of killing the entire
worker, meaning that a whole bunch of tests get aborted.

* chore(spread)!: rename "ubuntu:22.04" to "ubuntu@22.04"

This is temporary while we fix the issue related to the build plan rejecting
the ':' notation.

* build: pin craft-platforms to fixed upstream version

* fix: pass the correct base to the image service

* fix: pass the correct base to the package service

* chore(spread): run some tests last

These tests manipulate the docker snap itself, so let's run them last so that
they don't leave a broken system for the next tests.

* chore(spread): run 'pack' instead of 'prime'

The test itself just needs to build, it doesn't matter whether it's priming or
packing - and there's currently an unrelated bug with 'prime'.

* fix(test): use host architecture for test project

* fix: pr suggestion

Co-authored-by: Tiago Nobrega <tiago.nobrega@canonical.com>

* build(deps): update to craft-application 5.0.1

* chore!(schema): update schema for craft-app 5.0.1

The schema is now stricter than what the application supports for `platform`
entries.

* chore: drop '--no-hashes' in snap

* chore: lower priority of expressjs doc test

There is something wrong with the test that needs to be investigated: it times
out when packing the rock, and then the 'restore' script times out when cleaning
the lxd instance.

* chore: re-add hash in requirements-noble.txt

---------

Co-authored-by: Alex Lowe <alex.lowe@canonical.com>
Co-authored-by: Tiago Nobrega <tiago.nobrega@canonical.com>