feat(12-factor): extensions are updated to let charms use non-root users (#833)
As of Juju 3.6 Charms can have non-root users. We want to make use of the non-root charm containers. We had some issues with some file/folder permissions because we are creating/modifying some files from the charm code.

This PR fixes these permission issues by creating the folders inside the rockcraft.yaml file and setting the user as _daemon_. Adds a new part into all 12 Factor extensions called logging. This part creates the folders necessary for logging and sets their user to _daemon_ so that the non-root charms can interact with them.

For Flask and Django extensions, it also sets the gunicorn.conf.py file and its parent folder to _daemon_ user.
Again for Flask and Django extensions, it creates the /var/log/django and /var/log/flask folders and sets their user to _daemon_.