refactor: consolidate base-specific knowledge
Instead of having the information about which base behaves which way with
regards to support, pebble, plugins, etc, have a single "bases" module that
serves as an authoritative source for all base-related knowledge. The idea
is that someone can look at this single module and have a full view of how
Rockcraft's behavior changes depending on the project's base.