test(spread): add spread tests for Python plugin v2
The tests cover the basic cases of whether the Python interpreter is provided
by the part itself or an earlier one, whether it comes from Chisel slices, and
both 'bare' and Ubuntu bases.

Also bump up the number of spread workers a bit; this is tangentially related
to the task but the spread tests are taking a long time nowadays.