test(spread): fix python3.6 test (#1021)
The actual purpose of the test is to check that a rock that is built using the
Python plugin but without a 'python3' symlink works correctly if
PARTS_PYTHON_INTERPRETER is correctly set.

The test checked this by using python3.6 from the deadsnakes ppa, whose dependency
chain did *not* include 'python3-minimal' (I think), which is the package that
provides 'python3'.

Since python3.6 is gone from deadsnakes, update the test to be more explicit about
the intention by using the default Python interpreter for the base but removing
'python3' out "manually".